backend/Prowritesolutions/requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
bcrypt==4.0.1
PyJWT==2.8.0
requests==2.31.0
python-dotenv==1.0.0
cryptography==41.0.4
reportlab==4.0.4
Pillow==10.0.0
openai==0.28.0
gunicorn==21.2.0
sendgrid==6.10.0